requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0
beautifulsoup4==4.12.2
Pillow==10.2.0
python-multipart==0.0.9